Try on moving to render package for printing log entries
